Implemented KMP

+++ include/algorithms/string.hpp
@@ -17,10 +17,10 @@ namespace sangi {
     std::vector<std::string> LongestCommonSubstring(const std::string a, const std::string b);
 
     // Return indices where pattern appears in text using brute force search
-    std::vector<int> BruteForcePatternSearch(const std::string text, const std::string pattern);
+    std::vector<size_t> BruteForcePatternSearch(const std::string text, const std::string pattern);
 
     // Return indices where pattern appears in text using the KMP algorithm
-    std::vector<int> KMPSearch(const std::string text, const std::string pattern);
+    std::vector<size_t> KMPSearch(const std::string text, const std::string pattern);
 
 } // namespace sangi
